reworked auth on backend. method for current year report by salesman code

+++ backend/src/apis/reports-api.js
@@ -21,6 +21,9 @@ class reportApi {
 
     static getAllReports = async (req, res) => {
         try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             const reports = await ReportModel.find();
             res.status(200).send({apiStatus: true, message: "All reports fetched", data: reports})
         } catch (e) {
@@ -30,6 +33,9 @@ class reportApi {
 
     static getAllReportsByYear = async (req, res) => {
         try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             const reports = await ReportModel.find({year: req.params.year});
             res.status(200).send({
                 apiStatus: true,
@@ -43,7 +49,29 @@ class reportApi {
 
     static getAllReportsBySalesmanCode = async (req, res) => {
         try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             const data = await ReportModel.find({salesman_code: req.params.code})
+            if (data != null)
+                res.status(200).send({
+                    apiStatus: true,
+                    message: "Report record by code " + req.params.code + " was found",
+                    data: data
+                })
+            else
+                res.status(404).send({apiStatus: false, message: "Report record not found"})
+        } catch (e) {
+            res.status(500).send({apiStatus: false, message: e.message, data: e})
+        }
+    }
+
+    static getReportBySalesmanCodeForCurrentYear = async (req, res) => {
+        try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
+            const data = await ReportModel.find({salesman_code: req.params.code, year: new Date().getFullYear()})
             if (data != null)
                 res.status(200).send({
                     apiStatus: true,
@@ -59,6 +87,9 @@ class reportApi {
 
     static getReportById = async (req, res) => {
         try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1 && req.session.user.role !== 2) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             const data = await ReportModel.findById(req.params.id)
             if (data == null) {
                 return res.status(404).send({
@@ -77,6 +108,9 @@ class reportApi {
     }
     static updateReport = async (req, res) => {
         try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             if (req.params.id === undefined || req.body === undefined || req.params.id === "" || req.params.body === "") {
                 return res.status(400).send({apiStatus: false, message: "Invalid request parameters"});
             }
@@ -94,6 +128,9 @@ class reportApi {
 
     static submitReportByCEO = async (req, res) => {
         try {
+            if (req.session.user.role !== 0) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             if (req.params.id === undefined || req.params.id === "") {
                 return res.status(400).send({apiStatus: false, message: "Invalid request parameters"});
             }
@@ -111,6 +148,9 @@ class reportApi {
 
     static submitAllReportsByCEO = async (req, res) => {
         try {
+            if (req.session.user.role !== 0) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             let found = await ReportModel.find()
             if (found == null || found.length === 0) {
                 return res.status(404).send({message: "There are no Reports"});
@@ -130,6 +170,9 @@ class reportApi {
 
     static unSubmitReportByCEO = async (req, res) => {
         try {
+            if (req.session.user.role !== 0) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
             if (req.params.id === undefined || req.params.id === "") {
                 return res.status(400).send({apiStatus: false, message: "Invalid request parameters"});
             }
@@ -147,6 +190,10 @@ class reportApi {
 
     static unSubmitAllReportsByCEO = async (req, res) => {
         try {
+            if (req.session.user.role !== 0) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
+
             let found = await ReportModel.find()
             if (found == null || found.length === 0) {
                 return res.status(404).send({apiStatus: false, message: "There are no Reports"});
@@ -164,7 +211,7 @@ class reportApi {
         }
     }
 
-    static confirmationReverseWithIdsArrayByCEO = async (req, res) => {
+    static confirmationReverseWithIdsArray = async (req, res) => {
         try {
             const idArray = req.body.ids;
 
@@ -178,7 +225,18 @@ class reportApi {
                 return res.status(404).json({ error: 'No reports found for the provided IDs' });
             }
 
-            const result = await reportService.confirmationReverseWithIdsArrayBbyCEO(reports);
+            let result;
+
+            switch (req.session.user.role) {
+                case 0:
+                    result = await reportService.confirmationReverseWithIdsArrayBbyCEO(reports);
+                    break;
+                case 1:
+                    result = await reportService.confirmationReverseWithIdsArrayBbyHR(reports);
+                    break;
+                default:
+                    return res.status(403).json({ error: 'Unauthorized role' });
+            }
 
             res.status(200).json({ message: 'Reports processed successfully', result });
         } catch (error) {
@@ -187,55 +245,65 @@ class reportApi {
         }
     };
 
-    static confirmationPairsArrayByCEO = async (req, res) => {
-        try {
-            const pairsArray = req.body.pairs;
 
-            if (!Array.isArray(pairsArray) || pairsArray.length === 0) {
-                return res.status(400).json({ error: 'Invalid or empty pairs array' });
-            }
-
-            const ids = pairsArray.map((entry) => entry._id);
-            const reports = await ReportModel.find({ _id: { $in: ids } });
-
-            if (!reports || reports.length === 0) {
-                return res.status(404).json({ error: 'No reports found for the provided IDs' });
-            }
-
-            const result = await reportService.confirmationPairsArrayByCEO(reports, pairsArray);
-
-            res.status(200).json({ message: 'Reports processed successfully', result });
-        } catch (error) {
-            console.error('Error processing reports:', error);
-            res.status(500).json({ error: 'Internal server error' });
-        }
-    };
+    // static confirmationPairsArrayByCEO = async (req, res) => {
+    //     try {
+    //         const pairsArray = req.body.pairs;
+    //
+    //         if (!Array.isArray(pairsArray) || pairsArray.length === 0) {
+    //             return res.status(400).json({ error: 'Invalid or empty pairs array' });
+    //         }
+    //
+    //         const ids = pairsArray.map((entry) => entry._id);
+    //         const reports = await ReportModel.find({ _id: { $in: ids } });
+    //
+    //         if (!reports || reports.length === 0) {
+    //             return res.status(404).json({ error: 'No reports found for the provided IDs' });
+    //         }
+    //
+    //         const result = await reportService.confirmationPairsArrayByCEO(reports, pairsArray);
+    //
+    //         res.status(200).json({ message: 'Reports processed successfully', result });
+    //     } catch (error) {
+    //         console.error('Error processing reports:', error);
+    //         res.status(500).json({ error: 'Internal server error' });
+    //     }
+    // };
 
 
     static deleteAllReports = async (req, res) => {
         try {
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
+
             const reports = await ReportModel.deleteMany();
-            res.status(200).send({apiStatus: true, message: "All reports deleted", data: reports})
+            res.status(200).send({ apiStatus: true, message: "All reports deleted", data: reports });
         } catch (e) {
-            res.status(500).send({apiStatus: false, message: e.message, data: e})
+            res.status(500).send({ apiStatus: false, message: e.message, data: e });
         }
-    }
+    };
 
     static deleteReport = async (req, res) => {
         try {
-            const report = await ReportModel.findByIdAndDelete(req.params.id)
+            if (req.session.user.role !== 0 && req.session.user.role !== 1) {
+                return res.status(403).send({ apiStatus: false, message: "Unauthorized role" });
+            }
+
+            const report = await ReportModel.findByIdAndDelete(req.params.id);
             if (!report) {
-                return res.status(404).send({apiStatus: false, message: "Report not found"});
+                return res.status(404).send({ apiStatus: false, message: "Report not found" });
             }
 
             return res.status(200).send({
                 apiStatus: true,
                 message: "Report deleted",
-            })
+            });
         } catch (e) {
-            return res.status(500).send({apiStatus: true, message: e.message, data: e})
+            return res.status(500).send({ apiStatus: false, message: e.message, data: e });
         }
-    }
+    };
+
 
 }
